Restructuring Code
